# spec/api-docs/bolsas-pacientes/openapi-specification.yaml
openapi: 3.0.3
info:
  title: CENATE - Módulo Bolsas de Pacientes API
  description: |
    API REST para la gestión del módulo Bolsas de Pacientes del Sistema de Telemedicina CENATE (EsSalud Perú).

    ## ¿Qué son las Bolsas de Pacientes?

    Una **Bolsa** es un conjunto de pacientes agrupados por criterios específicos (diagnóstico, especialidad, urgencia)
    que requieren atención coordinada a través de CENATE. Coordina atenciones médicas remotas para **4.6M asegurados**
    a través de **414 IPRESS**.

    ## Características principales

    - ✅ Importación masiva desde Excel (v1.11.0 con auto-detección inteligente)
    - ✅ Gestión completa CRUD de solicitudes
    - ✅ Auto-enriquecimiento de datos desde BD (IPRESS, RED, Estados)
    - ✅ Soft delete con auditoría completa
    - ✅ Estados de gestión de citas integrados
    - ✅ Asignación de admisionistas y gestores
    - ✅ Filtrado y búsqueda avanzada

    ## Seguridad

    Todos los endpoints requieren autenticación JWT Bearer token.
    Algunos endpoints adicionales requieren permisos MBAC (Module-Based Access Control).

    ## Base de Datos

    - **Host:** 10.0.89.13:5432
    - **Database:** maestro_cenate
    - **Principales tablas:** dim_solicitud_bolsa, dim_tipos_bolsas, bolsa_107_item, dim_ipress, dim_red

  version: 1.11.0
  contact:
    name: Ing. Styp Canto Rondón
    email: stypcanto@essalud.gob.pe
  license:
    name: EsSalud Perú - Uso Interno
servers:
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /api/bolsas/solicitudes/importar:
    post:
      tags:
        - Solicitudes de Bolsa
      summary: Importar solicitudes desde Excel
      description: |
        Importa solicitudes de bolsa desde un archivo Excel con validación y enriquecimiento automático.

        **Features v1.11.0:**
        - ✅ Auto-detección inteligente de tipo de bolsa desde nombre de archivo
        - ✅ Auto-selección de especialidad/servicio con fuzzy matching
        - ✅ Validación de estructura Excel sin headers
        - ✅ Enriquecimiento automático de datos desde BD (LEFT JOINs)
        - ✅ Cálculo automático de edad

        **Columnas requeridas del Excel:**
        1. DNI (8 dígitos)
        2. Código Adscripción (IPRESS)
        3. Fecha preferida no atendida (opcional, formato: DD/MM/YYYY)
        4. Tipo documento (opcional: DNI, CE, etc.)
        5. Fecha nacimiento (opcional: DD/MM/YYYY)
        6. Sexo (opcional: M/F)
        7. Teléfono (opcional)
        8. Email (opcional)
        9. Tipo cita (opcional)
        10. Especialidad (opcional)

        **Proceso:**
        1. Validación estructura archivo
        2. Lectura fila por fila (sin header)
        3. Enriquecimiento: busca paciente, IPRESS, RED en BD
        4. Cálculo edad si hay fecha nacimiento
        5. Inserción en dim_solicitud_bolsa vía SP
        6. Retorna estadísticas (total, OK, errores)

      operationId: importarSolicitudesDesdeExcel
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - idBolsa
                - idServicio
              properties:
                file:
                  type: string
                  format: binary
                  description: Archivo Excel (.xls, .xlsx) con 10 columnas de datos
                idBolsa:
                  type: integer
                  format: int64
                  description: ID del tipo de bolsa (PASO 1 - de dim_tipos_bolsas)
                  example: 4
                idServicio:
                  type: integer
                  format: int64
                  description: ID del servicio/especialidad (PASO 2 - de tabla servicios)
                  example: 89
                usuarioCarga:
                  type: string
                  description: Usuario que realiza la carga
                  default: admin
                  example: admin
      responses:
        '200':
          description: Importación completada (puede tener errores parciales)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  totalFilas:
                    type: integer
                    example: 150
                    description: Total de filas procesadas del Excel
                  filasOk:
                    type: integer
                    example: 145
                    description: Filas importadas exitosamente
                  filasError:
                    type: integer
                    example: 5
                    description: Filas con errores
                  errores:
                    type: array
                    description: Detalle de errores por fila
                    items:
                      type: object
                      properties:
                        fila:
                          type: integer
                          example: 23
                        error:
                          type: string
                          example: "DNI inválido: debe tener 8 dígitos"
                  message:
                    type: string
                    example: "Importación completada: 145 registros insertados, 5 errores"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2026-01-27T10:30:00Z"
              examples:
                success:
                  value:
                    status: 200
                    totalFilas: 150
                    filasOk: 150
                    filasError: 0
                    errores: []
                    message: "Importación completada exitosamente"
                    timestamp: "2026-01-27T10:30:00Z"
                partial:
                  value:
                    status: 200
                    totalFilas: 150
                    filasOk: 145
                    filasError: 5
                    errores:
                      - fila: 12
                        error: "DNI inválido: debe tener 8 dígitos"
                      - fila: 35
                        error: "Código IPRESS no encontrado en BD"
                    message: "Importación parcial: 145 OK, 5 errores"
                    timestamp: "2026-01-27T10:30:00Z"
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidFile:
                  value:
                    status: 400
                    error: "INVALID_FILE"
                    message: "El archivo debe ser formato Excel (.xls o .xlsx)"
                    timestamp: "2026-01-27T10:30:00Z"
                missingColumns:
                  value:
                    status: 400
                    error: "INVALID_STRUCTURE"
                    message: "El Excel debe tener exactamente 10 columnas"
                    timestamp: "2026-01-27T10:30:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/bolsas/solicitudes:
    get:
      tags:
        - Solicitudes de Bolsa
      summary: Listar todas las solicitudes activas
      description: |
        Obtiene todas las solicitudes de bolsa activas (no eliminadas lógicamente).

        **Datos retornados incluyen:**
        - Información del paciente (DNI, nombre, edad calculada, sexo, teléfono, email)
        - Tipo de bolsa (código y descripción)
        - IPRESS y RED enriquecidos desde BD
        - Estado de gestión de citas actual
        - Fechas de solicitud y actualización
      operationId: listarSolicitudes
      responses:
        '200':
          description: Lista de solicitudes
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SolicitudBolsaDTO'
                  message:
                    type: string
                    example: "Solicitudes obtenidas exitosamente"
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/bolsas/solicitudes/{id}:
    get:
      tags:
        - Solicitudes de Bolsa
      summary: Obtener una solicitud por ID
      operationId: obtenerSolicitudPorId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID de la solicitud
          example: 1523
      responses:
        '200':
          description: Solicitud encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    $ref: '#/components/schemas/SolicitudBolsaDTO'
                  message:
                    type: string
                    example: "Solicitud obtenida exitosamente"
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Solicitud no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    delete:
      tags:
        - Solicitudes de Bolsa
      summary: Eliminar una solicitud (soft delete)
      description: |
        Elimina lógicamente una solicitud (establece activo = false).
        La solicitud no se borra físicamente de la BD.
      operationId: eliminarSolicitud
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1523
      responses:
        '200':
          description: Solicitud eliminada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  mensaje:
                    type: string
                    example: "Solicitud eliminada exitosamente"
                  idSolicitud:
                    type: integer
                    format: int64
                    example: 1523
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/bolsas/solicitudes/borrar:
    post:
      tags:
        - Solicitudes de Bolsa
      summary: Eliminar múltiples solicitudes (soft delete en lote)
      description: |
        Elimina lógicamente múltiples solicitudes en una sola operación.
        Útil para limpiezas masivas desde la interfaz de administración.
      operationId: eliminarSolicitudesMultiples
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ids
              properties:
                ids:
                  type: array
                  items:
                    type: integer
                    format: int64
                  description: Lista de IDs de solicitudes a eliminar
                  example: [1523, 1524, 1525, 1526]
                  minItems: 1
      responses:
        '200':
          description: Solicitudes eliminadas exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  mensaje:
                    type: string
                    example: "4 solicitud(es) eliminada(s) exitosamente"
                  totalBorrados:
                    type: integer
                    example: 4
                  ids:
                    type: array
                    items:
                      type: integer
                      format: int64
                    example: [1523, 1524, 1525, 1526]
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 400
                error: "INVALID_REQUEST"
                message: "No se proporcionaron IDs para borrar"
                timestamp: "2026-01-27T10:30:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/bolsas/solicitudes/{id}/asignar:
    patch:
      tags:
        - Solicitudes de Bolsa
      summary: Asignar una gestora a una solicitud
      description: Asigna un usuario con rol de gestora a una solicitud específica
      operationId: asignarGestora
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1523
        - name: idGestora
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: ID de la usuaria gestora (de tabla usuarios)
          example: 45
      responses:
        '200':
          description: Gestora asignada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  mensaje:
                    type: string
                    example: "Gestora asignada exitosamente"
                  idSolicitud:
                    type: integer
                    format: int64
                    example: 1523
                  idGestora:
                    type: integer
                    format: int64
                    example: 45
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/bolsas/solicitudes/{id}/estado:
    patch:
      tags:
        - Solicitudes de Bolsa
      summary: Cambiar el estado de una solicitud
      description: |
        Actualiza el estado de gestión de citas de una solicitud.
        Estados disponibles están en tabla dim_estados_gestion_citas.
      operationId: cambiarEstadoSolicitud
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1523
        - name: nuevoEstadoId
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: ID del nuevo estado (de dim_estados_gestion_citas)
          example: 2
      responses:
        '200':
          description: Estado actualizado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  mensaje:
                    type: string
                    example: "Estado actualizado exitosamente"
                  idSolicitud:
                    type: integer
                    format: int64
                    example: 1523
                  nuevoEstadoId:
                    type: integer
                    format: int64
                    example: 2
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/bolsas/solicitudes/asegurados-nuevos:
    get:
      tags:
        - Solicitudes de Bolsa
      summary: Obtener asegurados nuevos detectados
      description: |
        Retorna lista de DNIs en solicitudes que no tienen correspondencia en tabla asegurados.
        Útil para detectar pacientes que necesitan ser registrados en el sistema principal.
      operationId: obtenerAseguradosNuevos
      responses:
        '200':
          description: Lista de asegurados nuevos
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  total:
                    type: integer
                    example: 15
                  asegurados:
                    type: array
                    items:
                      type: object
                      properties:
                        paciente_dni:
                          type: string
                          example: "12345678"
                        paciente_nombre:
                          type: string
                          example: "JUAN PEREZ GARCIA"
                        fecha_deteccion:
                          type: string
                          format: date-time
                  mensaje:
                    type: string
                    example: "Se encontraron 15 asegurados nuevos"
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ========================================
  # BOLSA 107
  # ========================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtenido del endpoint de autenticación.
        Formato: "Bearer {token}"

        El token debe incluirse en el header Authorization:
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        ```
  schemas:
...
    SolicitudBolsaDTO:
      type: object
      description: Solicitud de bolsa con datos enriquecidos desde BD
      properties:
        id_solicitud:
          type: integer
          format: int64
          example: 1523
        numero_solicitud:
          type: string
          example: "SOL-2026-001523"
        paciente_id:
          type: integer
          format: int64
          example: 456789
        paciente_nombre:
          type: string
          example: "JUAN PEREZ GARCIA"
        paciente_dni:
          type: string
          example: "12345678"
        especialidad:
          type: string
          example: "CARDIOLOGIA"
        fecha_preferida_no_atendida:
          type: string
          format: date
          example: "2026-01-15"
          nullable: true
        tipo_documento:
          type: string
          example: "DNI"
          nullable: true
        fecha_nacimiento:
          type: string
          format: date
          example: "1980-05-20"
          nullable: true
        paciente_sexo:
          type: string
          enum: [M, F]
          example: "M"
          nullable: true
        paciente_telefono:
          type: string
          example: "987654321"
          nullable: true
        paciente_email:
          type: string
          format: email
          example: "juan.perez@example.com"
          nullable: true
        paciente_edad:
          type: integer
          example: 45
          description: Calculado automáticamente desde fecha_nacimiento
          nullable: true
        codigo_ipress:
          type: string
          example: "00012345"
          nullable: true
        tipo_cita:
          type: string
          example: "PRIMERA VEZ"
          nullable: true
        id_bolsa:
          type: integer
          format: int64
          example: 4
        desc_tipo_bolsa:
          type: string
          example: "BOLSAS_EXPLOTADATOS"
          description: Enriquecido desde dim_tipos_bolsas
        id_servicio:
          type: integer
          format: int64
          example: 89
        codigo_adscripcion:
          type: string
          example: "00012345"
          nullable: true
        id_ipress:
          type: integer
          format: int64
          example: 123
          nullable: true
        estado:
          type: string
          example: "PENDIENTE"
        fecha_solicitud:
          type: string
          format: date-time
          example: "2026-01-27T10:30:00-05:00"
        fecha_actualizacion:
          type: string
          format: date-time
          example: "2026-01-27T14:20:00-05:00"
        estado_gestion_citas_id:
          type: integer
          format: int64
          example: 1
          nullable: true
        cod_estado_cita:
          type: string
          example: "PENDIENTE"
          nullable: true
        desc_estado_cita:
          type: string
          example: "Pendiente de programación"
          nullable: true
        desc_ipress:
          type: string
          example: "HOSPITAL NACIONAL REBAGLIATI"
          description: Enriquecido desde dim_ipress
          nullable: true
        desc_red:
          type: string
          example: "RED ASISTENCIAL REBAGLIATI"
          description: Enriquecido desde dim_red
          nullable: true
        activo:
          type: boolean
          example: true
          description: Soft delete flag
...
    ErrorResponse:
      type: object
      description: Estructura estándar de respuesta de error CENATE
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "INVALID_REQUEST"
          description: Código de error identificable
        message:
          type: string
          example: "El campo 'idBolsa' es obligatorio"
          description: Mensaje descriptivo del error
        validationErrors:
          type: object
          additionalProperties:
            type: string
          example:
            idBolsa: "El campo no puede estar vacío"
            idServicio: "Debe ser un número positivo"
          description: Errores de validación por campo (opcional)
        timestamp:
          type: string
          format: date-time
          example: "2026-01-27T10:30:00Z"

  # ========================================
  # RESPONSES REUTILIZABLES
  # ========================================
  responses:
    UnauthorizedError:
      description: Token JWT inválido, expirado o no proporcionado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 401
            error: "UNAUTHORIZED"
            message: "Token JWT inválido o expirado"
            timestamp: "2026-01-27T10:30:00Z"
    NotFoundError:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 404
            error: "NOT_FOUND"
            message: "Solicitud con ID 1523 no encontrada"
            timestamp: "2026-01-27T10:30:00Z"
...